Update AutoBuildPlayer to show cmd command when build is completed

+++ Assets/Utils_Global/Editor/AutoBuildPlayer.cs
@@ -1,19 +1,21 @@
-﻿using UnityEditor;
-using UnityEngine;
-using UnityEditor.Build.Reporting;
+﻿using System.Diagnostics;
 using System.IO;
+using UnityEditor;
+using UnityEditor.Build.Reporting;
+using UnityEngine;
 
 // Output the build size or a failure depending on BuildPlayer.
 public class AutoBuildPlayer : MonoBehaviour
 {
     /// <summary>
-    /// "C:\Program Files\Unity\Hub\Editor\2019.3.13f1\Editor\Unity.exe" -quit -batchmode -projectPath "C:\Users\ERIC\Workspaces\Unity_Workspace\Sandbox" -executeMethod AutoBuildPlayer.AutoBuild
+    /// "C:\Program Files\Unity\Hub\Editor\2019.1.14f1\Editor\Unity.exe" -quit -batchmode -projectPath "C:\Users\ERIC\Workspaces\Unity_Workspace\Sandbox" -executeMethod AutoBuildPlayer.AutoBuild
     /// </summary>
     private static string buildPath = Path.Combine(Directory.GetCurrentDirectory(), "Build");
-    
+
     [MenuItem("Build/AutoBuild")]
     public static void AutoBuild()
     {
+        System.Console.WriteLine("AUTO BUILD");
         CleanBuildDirectory();
 
         BuildPlayerOptions buildPlayerOptions = new BuildPlayerOptions();
@@ -27,26 +29,44 @@ public class AutoBuildPlayer : MonoBehaviour
 
         if (summary.result == BuildResult.Succeeded)
         {
-            Debug.Log("Build succeeded: " + summary.totalSize + " bytes");
+            UnityEngine.Debug.Log("Build succeeded: " + summary.totalSize + " bytes");
+            CmdCommand(@"/k echo Build succeeded");
         }
 
         if (summary.result == BuildResult.Failed)
         {
-            Debug.Log("Build failed");
+            UnityEngine.Debug.Log("Build failed");
+            CmdCommand(@"/k echo Build succeeded");
         }
     }
 
     private static void CleanBuildDirectory()
     {
-        //Remove previous builds
-        DirectoryInfo buildDirectory = new DirectoryInfo(buildPath);
-        foreach (FileInfo file in buildDirectory.EnumerateFiles())
+        //Remove previous builds        
+        DirectoryInfo buildDirectory = Directory.CreateDirectory(buildPath);
+        if (buildDirectory.Exists)
         {
-            file.Delete();
+            foreach (FileInfo file in buildDirectory.EnumerateFiles())
+            {
+                file.Delete();
+            }
+            foreach (DirectoryInfo dir in buildDirectory.EnumerateDirectories())
+            {
+                dir.Delete(true);
+            }
         }
-        foreach (DirectoryInfo dir in buildDirectory.EnumerateDirectories())
+    }
+
+    public static void CmdCommand(string command)
+    {
+        var processInfo = new ProcessStartInfo("cmd.exe", command)
         {
-            dir.Delete(true);
-        }
+            CreateNoWindow = false,
+            UseShellExecute = true
+        };
+
+        var process = Process.Start(processInfo);
+
+        process.WaitForExit();
     }
 }
